Project/Software/Server/requirements.txt
Django
pyserial
numpy
matplotlib

